@@ -1,13 +1,26 @@
 package sperka.pl.bookcase.server.auth;
 
 import io.jsonwebtoken.security.Keys;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import org.eclipse.microprofile.config.inject.ConfigProperty;
 
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
 
+@ApplicationScoped
 public class SigningKey {
-    public static Key getKey() {
-        final String key = "E(H+MbQeThWmZq4t7w9z$C&F)J@NcRfU";
-        return Keys.hmacShaKeyFor(key.getBytes( StandardCharsets.UTF_8));
+    private final String key;
+
+    @Inject
+    public SigningKey( @ConfigProperty( name = "application.password.secret", defaultValue = "" ) String key ) {
+        this.key = key;
+        if ( key.length() < 32 ) {
+            throw new IllegalArgumentException( "application.password.secret needs to be at least 32 characters long" );
+        }
+    }
+
+    public Key getKey() {
+        return Keys.hmacShaKeyFor( key.getBytes( StandardCharsets.UTF_8 ) );
     }
 }
