@@ -1,5 +1,7 @@
 package sperka.pl.bookcase.server.dto;
 
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,7 +14,11 @@ import java.util.List;
 @Getter
 @Setter
 public class CreateUserRequestDto {
+    @NotEmpty( message = "user.error.name-must-not-be-empty" )
     private String name;
-    private String password;
+    @NotNull( message = "user.error.roles-must-not-be-empty" )
     private List< String > roles;
+    @NotEmpty( message = "user.error.email-must-not-be-empty" )
+    private String email;
+    private String locale;
 }
