@@ -1,12 +1,10 @@
 package sperka.pl.bookcase.server.repository.impl;
 
+import jakarta.enterprise.context.ApplicationScoped;
 import sperka.pl.bookcase.server.dto.LogFilterDto;
 import sperka.pl.bookcase.server.entity.Log;
 import sperka.pl.bookcase.server.repository.LogRepository;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
 import java.util.List;
 
 @ApplicationScoped
@@ -22,9 +20,7 @@ public class LogRepositoryImpl extends BasicRepositoryImpl< Log > implements Log
         var rootEntry = criteriaQuery.from( Log.class );
 
         criteriaQuery.select( rootEntry );
-//        Predicate whereCriteria = getFilterPredicate( filters, rootEntry );
 
-//        criteriaQuery.where( whereCriteria );
         if ( filters.getSortField() != null && !filters.getSortField().isEmpty() ) {
             if ( "desc".equals( filters.getSortDirection() ) ) {
                 criteriaQuery.orderBy( criteriaBuilder.desc( rootEntry.get( filters.getSortField() ) ) );
@@ -47,19 +43,9 @@ public class LogRepositoryImpl extends BasicRepositoryImpl< Log > implements Log
         var rootEntry = criteriaQuery.from( Log.class );
 
         criteriaQuery.select( criteriaBuilder.count( rootEntry ) );
-//        Predicate whereCriteria = getFilterPredicate( filters, rootEntry );
-
-//        criteriaQuery.where( whereCriteria );
 
         var query = entityManager.createQuery( criteriaQuery );
         return query.getSingleResult();
-    }
-
-    private Predicate getFilterPredicate( LogFilterDto filters, Root< Log > rootEntry ) {
-        var criteriaBuilder = entityManager.getCriteriaBuilder();
-        var whereCriteria = criteriaBuilder.ge( rootEntry.get( "id" ), 0 );
-
-        return whereCriteria;
     }
 
 }
