@@ -1,9 +1,8 @@
 package sperka.pl.bookcase.server.repository.impl;
 
+import jakarta.enterprise.context.ApplicationScoped;
 import sperka.pl.bookcase.server.entity.User;
 import sperka.pl.bookcase.server.repository.UserRepository;
-
-import javax.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
 public class UserRepositoryImpl extends BasicRepositoryImpl< User > implements UserRepository {
@@ -14,7 +13,19 @@ public class UserRepositoryImpl extends BasicRepositoryImpl< User > implements U
     @Override
     public User getUserByUsername( String username ) {
         return entityManager.createQuery( "select u from User u where u.name like :name", User.class )
-                .setParameter( "name", username ).getSingleResult();
+                .setParameter( "name", username ).getResultStream().findFirst().orElse( null );
+    }
+
+    @Override
+    public User getUserByResetPasswordToken( String resetToken ) {
+        return entityManager.createQuery( "select u from User u where u.resetPasswordToken = :token", User.class )
+                .setParameter( "token", resetToken ).getResultStream().findFirst().orElse( null );
+    }
+
+    @Override
+    public User getUserByEmail( String email ) {
+        return entityManager.createQuery( "select u from User u where u.email = :email", User.class )
+                .setParameter( "email", email ).getResultStream().findFirst().orElse( null );
     }
 
     @Override
