@@ -3,6 +3,8 @@ package sperka.pl.bookcase.server.service.impl;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
 import io.quarkus.scheduler.Scheduled;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import sperka.pl.bookcase.server.auth.SigningKey;
 import sperka.pl.bookcase.server.entity.JwtBlacklist;
@@ -12,8 +14,6 @@ import sperka.pl.bookcase.server.repository.UserRepository;
 import sperka.pl.bookcase.server.service.AuthService;
 import sperka.pl.bookcase.server.service.LogService;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.transaction.Transactional;
 import java.sql.Date;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
@@ -25,14 +25,17 @@ public class AuthServiceImpl implements AuthService {
     private final JwtBlacklistRepository jwtBlacklistRepository;
     private final UserRepository userRepository;
     private final LogService logService;
+    private final SigningKey signingKey;
 
-    public AuthServiceImpl( JwtBlacklistRepository jwtBlacklistRepository, UserRepository userRepository, LogService logService ) {
+    public AuthServiceImpl( JwtBlacklistRepository jwtBlacklistRepository, UserRepository userRepository, LogService logService, SigningKey signingKey ) {
         this.jwtBlacklistRepository = jwtBlacklistRepository;
         this.userRepository = userRepository;
         this.logService = logService;
+        this.signingKey = signingKey;
     }
 
     @Override
+    @Transactional
     public String loginUser( String username, String password ) {
         if ( username == null || password == null ) {
             return null;
@@ -40,6 +43,11 @@ public class AuthServiceImpl implements AuthService {
 
         var user = userRepository.getUserByUsername( username );
         if ( user != null && user.isPasswordMatching( password ) ) {
+            if ( user.getResetPasswordToken() != null ) {
+                // If someone logged in successfully, revoke password change token == they know password
+                revokeResetPasswordToken( user );
+            }
+
             if ( user.getActive() ) {
                 logService.add( "User '" + username + "' authenticated successfully", "system" );
                 return getJwtToken( user );
@@ -71,7 +79,7 @@ public class AuthServiceImpl implements AuthService {
 
     @Override
     public String refresh( String token ) {
-        var jwt = Jwts.parserBuilder().setSigningKey( SigningKey.getKey() ).build().parseClaimsJws( token );
+        var jwt = Jwts.parserBuilder().setSigningKey( signingKey.getKey() ).build().parseClaimsJws( token );
         var user = userRepository.getUserByUsername( jwt.getBody().getSubject() );
         if ( user != null && user.getActive() ) {
             if ( jwt.getBody().getExpiration().toInstant().isBefore( Instant.now().plus( 5, ChronoUnit.MINUTES ) ) ) {
@@ -85,15 +93,20 @@ public class AuthServiceImpl implements AuthService {
         return null;
     }
 
-    private static String getJwtToken( User user ) {
+    private String getJwtToken( User user ) {
         return Jwts.builder()
                 .setIssuer( "Bookshelf" )
                 .setSubject( user.getName() )
                 .claim( "roles", user.getRoles() )
                 .setIssuedAt( Date.from( Instant.now() ) )
                 .setExpiration( Date.from( Instant.now().plusMillis( 1000 * 60 * 60 ) ) )
-                .signWith( SigningKey.getKey() )
+                .signWith( signingKey.getKey() )
                 .compact();
+    }
+
+    private void revokeResetPasswordToken( User user ) {
+        user.setResetPasswordToken( null );
+        userRepository.save( user );
     }
 
     @Scheduled( every = "10m" )
