@@ -14,17 +14,20 @@ import { useApi } from "../../Services/GenericServiceHook";
 import { useEffect, useState } from "react";
 import { addUserParams, editUserParams, getUserParams } from "../../Services/UserApi";
 import { useUserContext } from "../../UserContextProvider";
-import { FormattedMessage } from "react-intl";
+import { FormattedMessage, useIntl } from "react-intl";
 
 const AddEditUser = props => {
+    const intl = useIntl();
     const { user } = useUserContext();
     const saveApi = useApi()
     const getUserApi = useApi()
     const [ formData, setFormData ] = useState( { id: null, name: "", roles: "", password: "", active: true } );
     const [ selfEditing, setSelfEditing ] = useState( false );
+    const [ validationErrors, setValidationErrors ] = useState( {} );
 
     useEffect( () => {
         if ( props.userId === parseInt( props.userId, 10 ) && props.open === true ) {
+            setValidationErrors( {} );
             getUserApi.request( getUserParams( props.userId ) );
         } else {
             resetForm();
@@ -51,21 +54,38 @@ const AddEditUser = props => {
     }, [ getUserApi.data ] )
 
     useEffect( () => {
-        if ( saveApi.error === "" && saveApi.loading === false && props.onClose ) {
+        if ( saveApi.data !== null ) {
+            if ( saveApi.data.status !== 200 ) {
+                saveApi.data.clone().json().then( json => {
+                    setValidationErrors( json );
+                } ).catch( () => {
+                    setValidationErrors( {} );
+                } );
+            } else {
+                setValidationErrors( {} );
+            }
+        }
+    }, [ saveApi.data ] )
+
+    useEffect( () => {
+        if ( saveApi.data !== null && saveApi.error === "" && saveApi.loading === false && props.onClose ) {
             props.onClose( true );
         }
     }, [ saveApi.error, saveApi.loading ] )
 
     const resetForm = () => {
-        setFormData( { id: null, name: "", roles: "", password: "", active: true } );
+        setValidationErrors( {} );
+        setFormData( { id: null, name: "", email: "", locale: "", roles: "", password: "", active: true } );
     }
 
     const submitUser = event => {
         event.preventDefault();
+        const roles = formData.roles === "" ? [] : formData.roles.split( ',' );
+
         if ( formData.id != null ) {
-            saveApi.request( editUserParams( { ...formData, roles: formData.roles.split( ',' ) } ) )
+            saveApi.request( editUserParams( { ...formData, roles: roles } ) )
         } else {
-            saveApi.request( addUserParams( { ...formData, roles: formData.roles.split( ',' ) } ) )
+            saveApi.request( addUserParams( { ...formData, roles: roles } ) )
         }
     }
 
@@ -81,6 +101,14 @@ const AddEditUser = props => {
 
     const onNameChange = event => {
         setFormData( prevState => ( { ...prevState, name: event.target.value } ) );
+    }
+
+    const onEmailChange = event => {
+        setFormData( prevState => ( { ...prevState, email: event.target.value } ) );
+    }
+
+    const onLocaleChange = event => {
+        setFormData( prevState => ( { ...prevState, locale: event.target.value } ) );
     }
 
     const onPasswordChange = event => {
@@ -99,16 +127,32 @@ const AddEditUser = props => {
         <Dialog className={ styles.dialog } open={ props.open }>
             <DialogTitle><FormattedMessage id="user-dialog.dialog-title"/></DialogTitle>
             <form className={ styles.form } onSubmit={ submitUser }>
-                { saveApi.error && <Alert severity={ "error" }>{ "Error: " + saveApi.error }</Alert> }
+                { saveApi.error && Object.keys( validationErrors ).length === 0 &&
+                    <Alert severity={ "error" }>{ "Error: " + saveApi.error }</Alert> }
                 <TextField name={ "name" } value={ formData.name } label={ <FormattedMessage id="user-dialog.name"/> }
-                           variant={ "standard" }
+                           variant={ "standard" } error={ 'name' in validationErrors }
+                           helperText={ 'name' in validationErrors && intl.formatMessage( { id: validationErrors[ 'name' ] } ) }
                            onChange={ onNameChange }/>
-                <TextField name={ "password" } value={ formData.password }
-                           label={ <FormattedMessage id="user-dialog.password"/> } variant={ "standard" }
-                           type={ "password" }
-                           onChange={ onPasswordChange }/>
+                <TextField name={ "email" } value={ formData.email }
+                           label={ <FormattedMessage id="user-dialog.email"/> }
+                           variant={ "standard" } error={ 'email' in validationErrors }
+                           helperText={ 'email' in validationErrors && intl.formatMessage( { id: validationErrors[ 'email' ] } ) }
+                           onChange={ onEmailChange }/>
+                <TextField name={ "locale" } value={ formData.locale }
+                           label={ <FormattedMessage id="user-dialog.locale"/> }
+                           variant={ "standard" } error={ 'locale' in validationErrors }
+                           helperText={ 'locale' in validationErrors && intl.formatMessage( { id: validationErrors[ 'locale' ] } ) }
+                           onChange={ onLocaleChange }/>
+                { props.userId !== null && <TextField name={ "password" } value={ formData.password }
+                                                      label={ <FormattedMessage id="user-dialog.password"/> }
+                                                      variant={ "standard" }
+                                                      type={ "password" } error={ 'password' in validationErrors }
+                                                      helperText={ 'password' in validationErrors && intl.formatMessage( { id: validationErrors[ 'password' ] } ) }
+                                                      onChange={ onPasswordChange }/> }
                 <TextField name={ "roles" } value={ formData.roles }
                            label={ <FormattedMessage id="user-dialog.roles"/> } variant={ "standard" }
+                           error={ 'roles' in validationErrors }
+                           helperText={ 'roles' in validationErrors && intl.formatMessage( { id: validationErrors[ 'roles' ] } ) }
                            onChange={ onRolesChange }/>
                 <FormControlLabel control={ <Checkbox/> }
                                   label={ <FormattedMessage id="user-dialog.active"/> } name={ "active" }
