@@ -30,6 +30,14 @@ const logoutParams = () => {
     return { url: "/api/auth/logout", method: "GET", data: null };
 }
 
+const getResetPasswordParams = ( token, data ) => {
+    return { url: "/api/user/reset-password/" + token, method: "POST", data: data };
+}
+
+const getRequestResetPasswordParams = ( data ) => {
+    return { url: "/api/user/request-reset-password", method: "POST", data: data };
+}
+
 export {
     getUserInfoParams,
     changeUserPasswordParams,
@@ -38,6 +46,8 @@ export {
     editUserParams,
     addUserParams,
     deleteUserParams,
-    logoutParams
+    logoutParams,
+    getResetPasswordParams,
+    getRequestResetPasswordParams
 };
 
